Cubic feedback

+++ packages/trpc/server/routers/viewer/teams/adminGetAll.schema.ts
@@ -1,8 +1,8 @@
 import { z } from "zod";
 
 export const ZAdminGetAllTeamsInputSchema = z.object({
-  limit: z.number().min(1).max(100).default(10),
-  offset: z.number().min(0).default(0),
+  limit: z.number().int().min(1).max(100).default(10),
+  offset: z.number().int().min(0).default(0),
   searchTerm: z.string().optional(),
   filters: z
     .array(
